Ajout du lien Modèle en disabled + refacto
